# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Black Tie VoIP Reseller Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/area-codes:
    get:
      operationId: adminGetAreaCodes
      tags: [dids]
      summary: Get all area codes
      responses:
        "200":
          description: List of area codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AreaCode"
    post:
      operationId: adminCreateAreaCode
      tags: [dids]
      summary: Create area code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAreaCodeRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AreaCode"
  /admin/area-codes/{id}:
    put:
      operationId: adminUpdateAreaCode
      tags: [dids]
      summary: Update area code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAreaCodeRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AreaCode"
    delete:
      operationId: adminDeleteAreaCode
      tags: [dids]
      summary: Delete area code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/dids:
    get:
      operationId: adminGetDids
      tags: [dids]
      summary: Get all DIDs with filters
      parameters:
        - name: areaCodeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: resellerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of DIDs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Did"
    post:
      operationId: adminCreateDid
      tags: [dids]
      summary: Create DID number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDidRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Did"
  /admin/dids/bulk:
    post:
      operationId: adminBulkCreateDids
      tags: [dids]
      summary: Bulk create DID numbers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateDidsRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/dids/{id}:
    put:
      operationId: adminUpdateDid
      tags: [dids]
      summary: Update DID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDidRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Did"
    delete:
      operationId: adminDeleteDid
      tags: [dids]
      summary: Delete DID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/dids/{id}/assign:
    post:
      operationId: adminAssignDid
      tags: [dids]
      summary: Assign DID to reseller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDidRequest"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Did"
  /admin/dids/{id}/unassign:
    post:
      operationId: adminUnassignDid
      tags: [dids]
      summary: Unassign DID from reseller
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Did"

  # ── Reseller endpoints ────────────────────────────────────────────────────────
...
  /reseller/dids/available:
    get:
      operationId: resellerGetAvailableDids
      tags: [dids]
      summary: Get available DIDs for an area code
      parameters:
        - name: areaCodeId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available DIDs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Did"
  /reseller/dids/{id}/request:
    post:
      operationId: resellerRequestDid
      tags: [dids]
      summary: Request/claim an available DID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DID assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Did"
...
  /admin/dids/import-gsheets:
    post:
      operationId: adminImportDidsFromSheets
      tags: [dids]
      summary: Import DIDs from a public Google Sheet (dry-run or actual)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportSheetsRequest"
      responses:
        "200":
          description: Import result or dry-run preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportSheetsResult"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    AreaCode:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        region:
          type: string
        province:
          type: string
        totalDids:
          type: integer
        availableDids:
          type: integer
        assignedDids:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, code, region, province, totalDids, availableDids, assignedDids, createdAt]
...
    CreateAreaCodeRequest:
      type: object
      properties:
        code:
          type: string
        region:
          type: string
        province:
          type: string
      required: [code, region, province]
    UpdateAreaCodeRequest:
      type: object
      properties:
        code:
          type: string
        region:
          type: string
        province:
          type: string
    Did:
      type: object
      properties:
        id:
          type: integer
        areaCodeId:
          type: integer
        areaCode:
          type: string
        region:
          type: string
        number:
          type: string
        status:
          type: string
          enum: [available, assigned, reserved]
        resellerId:
          type: integer
        resellerName:
          type: string
        assignedAt:
          type: string
          format: date-time
        reservedByOrderId:
          type: integer
          nullable: true
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, areaCodeId, number, status, createdAt]
    CreateDidRequest:
      type: object
      properties:
        areaCodeId:
          type: integer
        number:
          type: string
        notes:
          type: string
      required: [areaCodeId, number]
    BulkCreateDidsRequest:
      type: object
      properties:
        areaCodeId:
          type: integer
        numbers:
          type: array
          items:
            type: string
      required: [areaCodeId, numbers]
    UpdateDidRequest:
      type: object
      properties:
        notes:
          type: string
        status:
          type: string
          enum: [available, assigned, reserved]
    AssignDidRequest:
      type: object
      properties:
        resellerId:
          type: integer
      required: [resellerId]
...
    ImportSheetsRequest:
      type: object
      properties:
        url:
          type: string
        dryRun:
          type: boolean
      required: [url]
    ImportSheetsSampleRow:
      type: object
      properties:
        areaCode:
          type: string
        number:
          type: string
        region:
          type: string
        province:
          type: string
        notes:
          type: string
    ImportSheetsResult:
      type: object
      properties:
        ok:
          type: boolean
        totalRows:
          type: integer
        created:
          type: integer
        skipped:
          type: integer
        areaCodesCreated:
          type: integer
        sample:
          type: array
          items:
            $ref: "#/components/schemas/ImportSheetsSampleRow"
        headers:
          type: array
          items:
            type: string
      required: [ok, totalRows]
...